Resolve conflicts in batch_filter.rs (take main's SIMD version),
stream_executor.rs (re-apply pre-computed table_refs on main's base),
and Cross.toml (take main's version).

+++ crates/laminar-core/src/dag/tests.rs
@@ -3343,7 +3343,7 @@ fn test_operator_node_metrics_accumulate() {
     // Process 3 events
     for i in 0..3 {
         executor
-            .process_event(src_id, test_event(1000 + i, i as i64))
+            .process_event(src_id, test_event(1000 + i, i))
             .unwrap();
     }
